Added onClick for images
