+++ find_a/src/aboutUs.css
@@ -77,8 +77,33 @@
 }
 
 .avi:hover .loader {
+    pointer-events: none;
     transform: translate(-50%, calc(-50% - 1px)) scale(1);
     opacity: 1;
+}
+
+.overlay {
+    position: fixed;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    background-color: rgba(0, 0, 0, 0.75);
+    display: none;
+    justify-content: center;
+    align-items: center;
+    z-index: 3;
+}
+
+.circleClick {
+    width: 150px;
+    height: 150px;
+    border-radius: 50%;
+    background-color: #44ae72;
+}
+
+.overlay.overlayClick {
+    display: flex;
 }
 
 .avi {
@@ -344,4 +369,28 @@
 
 .jouji:hover::after {
     opacity: 1;
+}
+
+.overlay.aviClick .circleClick {
+    background-color: #44ae72;
+}
+
+.overlay.nehemiahClick .circleClick {
+    background-color: #3285f3;
+}
+
+.overlay.hamedClick .circleClick {
+    background-color: #333333;
+}
+
+.overlay.marionClick .circleClick {
+    background-color: #A65959;
+}
+
+.overlay.johnClick .circleClick {
+    background-color: #4E428A;
+}
+
+.overlay.joujiClick .circleClick {
+    background-color: #3E2723;
 }
